bug(api): fix multiple preview bugs + restructuring
