+++ apps/api/src/app/workflows-v2/usecases/extract-default-values-from-schema/extract-default-values-from-schema.usecase.ts
@@ -1,6 +1,6 @@
 import { JSONSchemaDto } from '@novu/shared';
 import { Injectable } from '@nestjs/common';
-import { ExtractDefaultValuesFromSchemaCommand } from './extract-default-values-from-schema-command';
+import { ExtractDefaultValuesFromSchemaCommand } from './extract-default-values-from-schema.command';
 
 @Injectable()
 export class ExtractDefaultValuesFromSchemaUsecase {
